requirements.txt
scikit-learn
fastapi
pandas
uvicorn
torch
torchvision
pillow
numpy
python-multipart
